requirements.txt
pandas
apache-airflow
apache-airflow-providers-postgres
apache-airflow-providers-common-sql
psycopg2-binary
asyncpg